+++ lib/ooxml_parser/pptx_parser/pptx_data/presentation.rb
@@ -7,12 +7,15 @@ require_relative 'presentation/presentation_theme'
 require_relative 'presentation/slide'
 require_relative 'presentation/slide_master'
 require_relative 'presentation/slide_masters_helper'
+require_relative 'presentation/slide_layout_file'
+require_relative 'presentation/slide_layouts_helper'
 require_relative 'presentation/slide_size'
 require_relative 'presentation/table_styles'
 module OoxmlParser
   # Basic class for all parsed pptx data
   class Presentation < CommonDocumentStructure
     include PresentationHelpers
+    include SlideLayoutsHelper
     include SlideMastersHelper
     attr_accessor :slides, :theme, :slide_size
     # @return [Relationships] relationships of presentation
@@ -23,13 +26,16 @@ module OoxmlParser
     attr_reader :comment_authors
     # @return [PresentationComments] comments of presentation
     attr_reader :comments
-    # @return [Array<SlideMaster] list of slide master
+    # @return [Array<SlideMaster>] list of slide master
     attr_reader :slide_masters
+    # @return [Array<SlideLayout>] list of slide layouts
+    attr_reader :slide_layouts
 
     def initialize(params = {})
       @slides = []
       @comments = []
       @slide_masters = []
+      @slide_layouts = []
       super
     end
 
@@ -62,6 +68,7 @@ module OoxmlParser
       end
       OOXMLDocumentObject.xmls_stack.pop
       @relationships = Relationships.new(parent: self).parse_file("#{OOXMLDocumentObject.path_to_folder}/ppt/_rels/presentation.xml.rels")
+      parse_slide_layouts
       parse_slide_masters
       self
     end