+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * Add `CodeQL` check in CI
 * Add parsing `Presentation#slide_master`
+* Add parsing `Presentation#slide_layouts`
 
 ### Changes